upstream enhancements:
Enhance LEMON to track which symbols actually carry semantic content. Output the list of symbols that do not carry content at the end of the report, but do not (yet) do anything else with the information.
Fix an unreachable branch associated with stack overflow in the LEMON-generated parser.
Performance improvements on the main loop of the LEMON-generated parser.
Add the %extra_context directive to lemon, as an alternative to %extra_argument. Use this to improve the performance of the parser.
Add the -dDIRECTORY command-line option to LEMON.
Lemon enhancements: (1) Do not allocate space for the 'error' non-terminal if it is not used. (2) Fix an off-by-one problem so that 'unsigned char' can be used for symbol numbers if the number of symbols is 256
Enhance LEMON to show precendence of symbols and all rules in the report that is generated in parallel to the parser.